requirements.txt
Pillow==9.4.0
requests==2.28.1
selenium==4.7.2
